move graphql to extension
